pyproject.toml
[project]
name = "vyapari-ai"
version = "0.1.0"
description = "AI Commerce Copilot for Bharat — Multi-agent system on Amazon Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.18",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "structlog>=24.0.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "api*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
